#283 put tests that require settle_bud method in src/bud/bud_settle directory
